--- a/src/llama_recipes/utils/flop_utils.py
+++ b/src/llama_recipes/utils/flop_utils.py
@@ -1,6 +1,8 @@
+from typing import Any, Dict, List, Optional, Union
+
 import torch
 from torch.utils.flop_counter import FlopCounterMode
-from typing import List, Any, Dict, Optional, Union
+
 
 class FlopMeasure(FlopCounterMode):
     """
@@ -22,6 +24,7 @@ class FlopMeasure(FlopCounterMode):
                 mod(batch)
                 flop_counter.stop_counting()
     """
+
     def __init__(
         self,
         mods: Optional[Union[torch.nn.Module, List[torch.nn.Module]]] = None,
@@ -33,18 +36,24 @@ class FlopMeasure(FlopCounterMode):
         super().__init__(mods, depth, display, custom_mapping)
         self.ready = False
         self.rank = rank
+
     def __enter__(self):
         self.ready = False
         super().__enter__()
         return self
+
     def get_total_flops(self):
         return super().get_total_flops()
+
     def get_table(self, depth=2):
         return super().get_table(depth)
+
     def __exit__(self, *args):
         self.ready = False
         if self.get_total_flops() == 0:
-            print("Warning: did not record any flops this time. Skipping the flop report")
+            print(
+                "Warning: did not record any flops this time. Skipping the flop report"
+            )
         else:
             self.stop_counting()
             if self.display:
@@ -57,10 +66,13 @@ class FlopMeasure(FlopCounterMode):
 
     def start_counting(self):
         self.ready = True
+
     def is_ready(self):
         return self.ready
+
     def stop_counting(self):
         self.ready = False
+
     def __torch_dispatch__(self, func, types, args=(), kwargs=None):
         # return the original output if not ready
         if not self.ready: