--- a/src/llama_recipes/utils/tflop_counter.py
+++ b/src/llama_recipes/utils/tflop_counter.py
@@ -445,19 +445,28 @@ class FlopCounterMode(TorchDispatchMode):
         return self
 
     def __exit__(self, *args):
-        self.stop_counting()
-        if self.display:
-            if self.rank is None or self.rank == 0:
-                print(self.get_table(self.depth))
+        if self.get_total_flops() == 0:
+            print("Warning: did not record any flops this time. Skipping the flop report")
+        else:
+            self.stop_counting()
+            if self.display:
+                if self.rank is None or self.rank == 0:
+                    print("exiting flop counter")
+                    print("self.flop_counts", self.flop_counts["Global"].values())
+                    print(self.get_table(self.depth))
         super().__exit__(*args)
     def start_counting(self):
         self.flop_counts.clear()
         self.ready = True
+        print("start_counting")
     def stop_counting(self):
         self.ready = False
+        print("stop_counting")
     def __torch_dispatch__(self, func, types, args=(), kwargs=None):
         if not self.ready:
+            print("not ready yet, stop_counting")
             return
+        print("in torch_dispatch now")
         kwargs = kwargs if kwargs else {}
         out = func(*args, **kwargs)
         func_packet = func._overloadpacket