# What does this PR do?

Added a feature that allow users to use pytorch profiler or flop_counter to measure the performance during fine-tuning. For pytorch profiler, use --use_profiler to turn on pytorch profiler, the default profiling schedule is  "wait_step, warmup_step, active_step = 1, 2, 3" , use --profiler_dir SOME_PATH to save the profiler output into the directory. For flop counter, use --flop_counter to turn on flop counter which will only measure the flop of 1 step and print the table. The default step that will be counted is 3, use --flop_counter_startpoint to change the startpoint, but please allows some warmup steps before counting.  --use_profiler and --flop_counter **can not** be turned on the same time to ensure the measurement is accurate. Please make sure the --max_train_steps is long enough to allow the measurement activities. 

## Feature/Issue validation/testing

Please describe the tests that you ran to verify your changes and relevant result summary. Provide instructions so it can be reproduced.
Please also list any relevant details for your test configuration.

- [x] Test only --use_profiler option 
Logs:
```
~/work/llama-recipes (feature/flop_counter)]$ torchrun --nnodes 1 --nproc_per_node 2  recipes/finetuning/finetuning.py --use_profiler --profiler_dir /home/kaiwu/work/finetune-output --max_train_step 8 --max_eval_step 2 --use_peft --peft_method lora  --model_name /home/kaiwu/work/llama2-7b --enable_fsdp --output_dir /home/kaiwu/work/finetune-output
[2024-04-29 16:31:12,562] torch.distributed.run: [WARNING] 
[2024-04-29 16:31:12,562] torch.distributed.run: [WARNING] *****************************************
[2024-04-29 16:31:12,562] torch.distributed.run: [WARNING] Setting OMP_NUM_THREADS environment variable for each process to be 1 in default, to avoid your system being overloaded, please further tune the variable for optimal performance in your application as needed. 
[2024-04-29 16:31:12,562] torch.distributed.run: [WARNING] *****************************************
/home/kaiwu/miniconda3/envs/llama/lib/python3.10/site-packages/transformers/utils/generic.py:485: UserWarning: torch.utils._pytree._register_pytree_node is deprecated. Please use torch.utils._pytree.register_pytree_node instead.
  _torch_pytree._register_pytree_node(
/home/kaiwu/miniconda3/envs/llama/lib/python3.10/site-packages/transformers/utils/generic.py:485: UserWarning: torch.utils._pytree._register_pytree_node is deprecated. Please use torch.utils._pytree.register_pytree_node instead.
  _torch_pytree._register_pytree_node(
/home/kaiwu/miniconda3/envs/llama/lib/python3.10/site-packages/transformers/utils/generic.py:342: UserWarning: torch.utils._pytree._register_pytree_node is deprecated. Please use torch.utils._pytree.register_pytree_node instead.
  _torch_pytree._register_pytree_node(
/home/kaiwu/miniconda3/envs/llama/lib/python3.10/site-packages/transformers/utils/generic.py:342: UserWarning: torch.utils._pytree._register_pytree_node is deprecated. Please use torch.utils._pytree.register_pytree_node instead.
  _torch_pytree._register_pytree_node(
Clearing GPU cache for all ranks
--> Running with torch dist debug set to detail
Loading checkpoint shards: 100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 2/2 [00:07<00:00,  3.58s/it]
--> Model /home/kaiwu/work/llama2-7b

--> /home/kaiwu/work/llama2-7b has 6738.415616 Million params

trainable params: 4,194,304 || all params: 6,742,609,920 || trainable%: 0.06220594176090199
bFloat16 enabled for mixed precision - using bfSixteen policy
Loading checkpoint shards: 100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 2/2 [00:06<00:00,  3.48s/it]
trainable params: 4,194,304 || all params: 6,742,609,920 || trainable%: 0.06220594176090199
--> applying fsdp activation checkpointing...
--> applying fsdp activation checkpointing...
Reusing dataset samsum (/home/kaiwu/.cache/huggingface/datasets/samsum/samsum/0.0.0/f1d7c6b7353e6de335d444e424dc002ef70d1277109031327bc9cc6af5d3d46e)
Parameter 'function'=<function get_preprocessed_samsum.<locals>.apply_prompt_template at 0x7f9dec107250> of the transform datasets.arrow_dataset.Dataset._map_single couldn't be hashed properly, a random hash was used instead. Make sure your transforms and parameters are serializable with pickle or dill for the dataset fingerprinting and caching to work. If you reuse this transform, the caching mechanism will consider it to be different from the previous calls and recompute everything. This warning is only showed once. Subsequent hashing failures won't be showed.
Loading cached processed dataset at /home/kaiwu/.cache/huggingface/datasets/samsum/samsum/0.0.0/f1d7c6b7353e6de335d444e424dc002ef70d1277109031327bc9cc6af5d3d46e/cache-1c80317fa3b1799d.arrow
Loading cached processed dataset at /home/kaiwu/.cache/huggingface/datasets/samsum/samsum/0.0.0/f1d7c6b7353e6de335d444e424dc002ef70d1277109031327bc9cc6af5d3d46e/cache-bdd640fb06671ad1.arrow
Reusing dataset samsum (/home/kaiwu/.cache/huggingface/datasets/samsum/samsum/0.0.0/f1d7c6b7353e6de335d444e424dc002ef70d1277109031327bc9cc6af5d3d46e)
Parameter 'function'=<function get_preprocessed_samsum.<locals>.apply_prompt_template at 0x7f941400b250> of the transform datasets.arrow_dataset.Dataset._map_single couldn't be hashed properly, a random hash was used instead. Make sure your transforms and parameters are serializable with pickle or dill for the dataset fingerprinting and caching to work. If you reuse this transform, the caching mechanism will consider it to be different from the previous calls and recompute everything. This warning is only showed once. Subsequent hashing failures won't be showed.
Loading cached processed dataset at /home/kaiwu/.cache/huggingface/datasets/samsum/samsum/0.0.0/f1d7c6b7353e6de335d444e424dc002ef70d1277109031327bc9cc6af5d3d46e/cache-1c80317fa3b1799d.arrow
Loading cached processed dataset at /home/kaiwu/.cache/huggingface/datasets/samsum/samsum/0.0.0/f1d7c6b7353e6de335d444e424dc002ef70d1277109031327bc9cc6af5d3d46e/cache-bdd640fb06671ad1.arrow
--> Training Set Length = 14732
Reusing dataset samsum (/home/kaiwu/.cache/huggingface/datasets/samsum/samsum/0.0.0/f1d7c6b7353e6de335d444e424dc002ef70d1277109031327bc9cc6af5d3d46e)
Loading cached processed dataset at /home/kaiwu/.cache/huggingface/datasets/samsum/samsum/0.0.0/f1d7c6b7353e6de335d444e424dc002ef70d1277109031327bc9cc6af5d3d46e/cache-3eb13b9046685257.arrow
Loading cached processed dataset at /home/kaiwu/.cache/huggingface/datasets/samsum/samsum/0.0.0/f1d7c6b7353e6de335d444e424dc002ef70d1277109031327bc9cc6af5d3d46e/cache-23b8c1e9392456de.arrow
Preprocessing dataset:  12%|██████████████████████████▍                                                                                                                                                                                           | 1817/14732 [00:00<00:02, 5967.38it/s]Reusing dataset samsum (/home/kaiwu/.cache/huggingface/datasets/samsum/samsum/0.0.0/f1d7c6b7353e6de335d444e424dc002ef70d1277109031327bc9cc6af5d3d46e)
Loading cached processed dataset at /home/kaiwu/.cache/huggingface/datasets/samsum/samsum/0.0.0/f1d7c6b7353e6de335d444e424dc002ef70d1277109031327bc9cc6af5d3d46e/cache-3eb13b9046685257.arrow
Loading cached processed dataset at /home/kaiwu/.cache/huggingface/datasets/samsum/samsum/0.0.0/f1d7c6b7353e6de335d444e424dc002ef70d1277109031327bc9cc6af5d3d46e/cache-23b8c1e9392456de.arrow
--> Validation Set Length = 818
Preprocessing dataset: 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 14732/14732 [00:02<00:00, 5900.94it/s]
Preprocessing dataset: 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 818/818 [00:00<00:00, 5699.45it/s]
Preprocessing dataset:  98%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████▏   | 14470/14732 [00:02<00:00, 5942.71it/s]/home/kaiwu/miniconda3/envs/llama/lib/python3.10/site-packages/torch/cuda/memory.py:330: FutureWarning: torch.cuda.reset_max_memory_allocated now calls torch.cuda.reset_peak_memory_stats, which resets /all/ peak memory stats.
  warnings.warn(
Training Epoch: 1:   0%|                                                                                                                                                                                                                                          | 0/96 [00:00<?, ?it/s]pytorch profiling is activated and results will be saved in /home/kaiwu/work/finetune-output
Preprocessing dataset: 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 14732/14732 [00:02<00:00, 5869.94it/s]
Preprocessing dataset: 100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 818/818 [00:00<00:00, 5462.13it/s]
/home/kaiwu/miniconda3/envs/llama/lib/python3.10/site-packages/torch/cuda/memory.py:330: FutureWarning: torch.cuda.reset_max_memory_allocated now calls torch.cuda.reset_peak_memory_stats, which resets /all/ peak memory stats.
  warnings.warn(
Training Epoch: 1:   0%|                                                                                                                                                                                                                                          | 0/96 [00:00<?, ?it/s]pytorch profiling is activated and results will be saved in /home/kaiwu/work/finetune-output
NCCL version 2.19.3+cuda12.1
Training Epoch: 1/3, step 1/96 completed (loss: 1.3550630807876587):   3%|█████▌                                                                                                                                                                          | 3/96 [00:15<07:11,  4.64s/it]STAGE:2024-04-29 16:31:45 648753:648753 ActivityProfilerController.cpp:314] Completed Stage: Warm Up
Training Epoch: 1/3, step 1/96 completed (loss: 1.487619400024414):   3%|█████▌                                                                                                                                                                           | 3/96 [00:14<07:03,  4.55s/it]STAGE:2024-04-29 16:31:45 648752:648752 ActivityProfilerController.cpp:314] Completed Stage: Warm Up
Training Epoch: 1/3, step 4/96 completed (loss: 1.346050500869751):   6%|███████████                                                                                                                                                                      | 6/96 [00:25<05:51,  3.90s/it]STAGE:2024-04-29 16:31:57 648753:648753 ActivityProfilerController.cpp:320] Completed Stage: Collection
STAGE:2024-04-29 16:31:57 648752:648752 ActivityProfilerController.cpp:320] Completed Stage: Collection
STAGE:2024-04-29 16:31:57 648753:648753 ActivityProfilerController.cpp:324] Completed Stage: Post Processing
STAGE:2024-04-29 16:31:57 648752:648752 ActivityProfilerController.cpp:324] Completed Stage: Post Processing
Training Epoch: 1/3, step 7/96 completed (loss: 1.2207212448120117):   8%|██████████████▋                                                                                                                                                                 | 8/96 [00:48<10:36,  7.23s/it]max training steps reached, stopping training, total train steps finished:  8
Training Epoch: 1/3, step 7/96 completed (loss: 1.2207212448120117):   8%|██████████████▋                                                                                                                                                                 | 8/96 [00:48<08:57,  6.11s/it]
Training Epoch: 1/3, step 7/96 completed (loss: 1.1940990686416626):   8%|██████████████▋                                                                                                                                                                 | 8/96 [00:49<09:02,  6.16s/it]
Max CUDA memory allocated was 23 GB
Max CUDA memory reserved was 27 GB
Peak active CUDA memory was 24 GB
CUDA Malloc retries : 0
CPU Total Peak Memory consumed during the train (max): 10 GB
evaluating Epoch:  10%|█████████████████████▌                                                                                                                                                                                                             | 2/21 [00:00<00:07,  2.60it/s]max eval steps reached, stopping evaluation, total_eval_steps:  2
evaluating Epoch:  10%|█████████████████████▌                                                                                                                                                                                                             | 2/21 [00:00<00:08,  2.21it/s]
evaluating Epoch:  10%|█████████████████████▌                                                                                                                                                                                                             | 2/21 [00:00<00:08,  2.31it/s]
 eval_ppl=tensor(1.1119, device='cuda:0') eval_epoch_loss=tensor(0.1061, device='cuda:0')
we are about to save the PEFT modules
/home/kaiwu/miniconda3/envs/llama/lib/python3.10/site-packages/peft/utils/save_and_load.py:154: UserWarning: Could not find a config file in /home/kaiwu/work/llama2-7b - will assume that the vocabulary was not modified.
  warnings.warn(
/home/kaiwu/miniconda3/envs/llama/lib/python3.10/site-packages/peft/utils/save_and_load.py:154: UserWarning: Could not find a config file in /home/kaiwu/work/llama2-7b - will assume that the vocabulary was not modified.
  warnings.warn(
PEFT modules are saved in /home/kaiwu/work/finetune-output directory
best eval loss on epoch 1 is 0.10606812685728073
Epoch 1: train_perplexity=1.1164, train_epoch_loss=0.1101, epoch time 50.1332068501506s
Key: avg_train_prep, Value: 1.1164158582687378
Key: avg_train_loss, Value: 0.11012342572212219
Key: avg_eval_prep, Value: 1.1118977069854736
Key: avg_eval_loss, Value: 0.10606812685728073
Key: avg_epoch_time, Value: 50.1332068501506
Key: avg_checkpoint_time, Value: 0.2338113421574235
```
- [x] Test only --flop_counter option
Logs:
```
 ~/work/llama-recipes (feature/flop_counter)]$ torchrun --nnodes 1 --nproc_per_node 2  recipes/finetuning/finetuning.py --flop_counter --flop_counter_start 4 --max_train_step 6 --max_eval_step 2 --use_peft --peft_method lora  --model_name /home/kaiwu/work/llama2-7b --enable_fsdp --output_dir /home/kaiwu/work/finetune-output
[2024-04-29 16:27:28,475] torch.distributed.run: [WARNING] 
[2024-04-29 16:27:28,475] torch.distributed.run: [WARNING] *****************************************
[2024-04-29 16:27:28,475] torch.distributed.run: [WARNING] Setting OMP_NUM_THREADS environment variable for each process to be 1 in default, to avoid your system being overloaded, please further tune the variable for optimal performance in your application as needed. 
[2024-04-29 16:27:28,475] torch.distributed.run: [WARNING] *****************************************
/home/kaiwu/miniconda3/envs/llama/lib/python3.10/site-packages/transformers/utils/generic.py:485: UserWarning: torch.utils._pytree._register_pytree_node is deprecated. Please use torch.utils._pytree.register_pytree_node instead.
  _torch_pytree._register_pytree_node(
/home/kaiwu/miniconda3/envs/llama/lib/python3.10/site-packages/transformers/utils/generic.py:485: UserWarning: torch.utils._pytree._register_pytree_node is deprecated. Please use torch.utils._pytree.register_pytree_node instead.
  _torch_pytree._register_pytree_node(
/home/kaiwu/miniconda3/envs/llama/lib/python3.10/site-packages/transformers/utils/generic.py:342: UserWarning: torch.utils._pytree._register_pytree_node is deprecated. Please use torch.utils._pytree.register_pytree_node instead.
  _torch_pytree._register_pytree_node(
/home/kaiwu/miniconda3/envs/llama/lib/python3.10/site-packages/transformers/utils/generic.py:342: UserWarning: torch.utils._pytree._register_pytree_node is deprecated. Please use torch.utils._pytree.register_pytree_node instead.
  _torch_pytree._register_pytree_node(
Clearing GPU cache for all ranks
--> Running with torch dist debug set to detail
Loading checkpoint shards: 100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 2/2 [00:07<00:00,  3.62s/it]
--> Model /home/kaiwu/work/llama2-7b

--> /home/kaiwu/work/llama2-7b has 6738.415616 Million params

trainable params: 4,194,304 || all params: 6,742,609,920 || trainable%: 0.06220594176090199
bFloat16 enabled for mixed precision - using bfSixteen policy
Loading checkpoint shards: 100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 2/2 [00:07<00:00,  3.65s/it]
trainable params: 4,194,304 || all params: 6,742,609,920 || trainable%: 0.06220594176090199
--> applying fsdp activation checkpointing...
Reusing dataset samsum (/home/kaiwu/.cache/huggingface/datasets/samsum/samsum/0.0.0/f1d7c6b7353e6de335d444e424dc002ef70d1277109031327bc9cc6af5d3d46e)
Parameter 'function'=<function get_preprocessed_samsum.<locals>.apply_prompt_template at 0x7f084c903250> of the transform datasets.arrow_dataset.Dataset._map_single couldn't be hashed properly, a random hash was used instead. Make sure your transforms and parameters are serializable with pickle or dill for the dataset fingerprinting and caching to work. If you reuse this transform, the caching mechanism will consider it to be different from the previous calls and recompute everything. This warning is only showed once. Subsequent hashing failures won't be showed.
Loading cached processed dataset at /home/kaiwu/.cache/huggingface/datasets/samsum/samsum/0.0.0/f1d7c6b7353e6de335d444e424dc002ef70d1277109031327bc9cc6af5d3d46e/cache-1c80317fa3b1799d.arrow
Loading cached processed dataset at /home/kaiwu/.cache/huggingface/datasets/samsum/samsum/0.0.0/f1d7c6b7353e6de335d444e424dc002ef70d1277109031327bc9cc6af5d3d46e/cache-bdd640fb06671ad1.arrow
--> Training Set Length = 14732
Reusing dataset samsum (/home/kaiwu/.cache/huggingface/datasets/samsum/samsum/0.0.0/f1d7c6b7353e6de335d444e424dc002ef70d1277109031327bc9cc6af5d3d46e)
Loading cached processed dataset at /home/kaiwu/.cache/huggingface/datasets/samsum/samsum/0.0.0/f1d7c6b7353e6de335d444e424dc002ef70d1277109031327bc9cc6af5d3d46e/cache-3eb13b9046685257.arrow
Loading cached processed dataset at /home/kaiwu/.cache/huggingface/datasets/samsum/samsum/0.0.0/f1d7c6b7353e6de335d444e424dc002ef70d1277109031327bc9cc6af5d3d46e/cache-23b8c1e9392456de.arrow
--> Validation Set Length = 818
…