@HamidShojanazeri 2024-04-29
@wukaixingxp can you pls add `tabulate` to the requirements as well.