@HamidShojanazeri 2024-04-29
@wukaixingxp can you pls also update relevant docs/ like under recipe/finetuning, that now we support flop count and profiling.