@wukaixingxp 2024-04-29
> @wukaixingxp can you pls also update relevant docs/ like under recipe/finetuning, that now we support flop count and profiling.

Updated docs in docs/single_gpu.md, docs/multi_gpu.md, recipes/finetuning/singlegpu_finetuning.md and recipes/finetuning/multigpu_finetuning.md