@wukaixingxp 2024-04-29
> for profiling it seems we need --max_train_step >7? shall we kind document that as well/ any specific logic for 7?

Updated more about this requirement in docs and error message